Refactored queries code into subsections.
